re-org some code
